🐛 FIX: Seperate admin view from app
