+++ frontend/src/pages/index.jsx
@@ -3,3 +3,4 @@ export { default as NotFound } from 'pages/NotFound'
 export { default as Contribute } from 'pages/Contribute'
 export { default as Contact } from 'pages/Contact'
 export { default as Login } from 'pages/Login'
+export { default as AdminView } from 'pages/AdminView'
